+++ src/data/sidelinks.tsx
@@ -195,19 +195,19 @@ export const sidelinks: SideLink[] = [
       {
         title: 'Online recruiment platform',
         label: '',
-        href: '/job-posting-list',
+        href: '/recruitment/web',
         icon: <IconFileText size={18} />,
       },
       {
-        title: 'Interview questions',
+        title: 'Interview Test',
         label: '',
-        href: '/job-list',
+        href: '/recruitment/interview-test',
         icon: <IconFileText size={18} />,
       },
       {
-        title: 'Job List',
+        title: 'Interview Question',
         label: '',
-        href: '/job-list',
+        href: '/recruitment/interview-question',
         icon: <IconFileText size={18} />,
       },
     ],
